requirements.txt
colorama==0.4.6
iniconfig==2.3.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
dotenv==1.2.1
PyQt5==5.15.11
